moved duplicate TimeInfo components to the common schedule.utils.tsx file
